requirements.txt
requests==2.13.0
Flask==0.12.2
elasticsearch==7.0.1
nltk==3.2.5
javalang==0.11.0
beautifulsoup4==4.7.1
PyGithub==1.43.7
PyYAML==5.1
